Real Time Mask Detection/requirements.txt
keras_nightly==2.5.0.dev2021032900
numpy==1.19.2
tensorflow==2.5.0
streamlit==0.73.0
opencv_python==4.4.0.46
keras==2.4.3
Pillow==8.3.1
